Refactor WriteImage: update image handling to use ImageHolder<float> for saveImage and ContourOverLay methods

+++ cpp_code/IO/WriteImage.hpp
@@ -9,13 +9,13 @@ class WriteImage
 {
     public:
         WriteImage();
-        void saveImage(const ImageHolder<uint8_t> &image);
+        void saveImage(const ImageHolder<float> &image);
 
-        void ContourOverLay(const Contour &contour, const ImageHolder<uint8_t> &image, const ImageProcessor &processor);
+        void ContourOverLay(const Contour &contour, const ImageHolder<float> &image, const ImageProcessor &processor);
 
         
         private:
-        void SaveAsPGM(const ImageHolder<uint8_t> &image);
+        void SaveAsPGM(const ImageHolder<float> &image);
         // void SaveAsPGM(const std::vector<std::vector<uint8_t>> &image);
 };
